feat(motr): add Motor WASM plugin for WebAuthn passkey recovery
