+++ libs/enclave/package.json
@@ -16,9 +16,10 @@
     "dist"
   ],
   "scripts": {
-    "build": "devbox run build && tsc --declaration --emitDeclarationOnly && tsc-alias",
+    "build": "devbox run build && tsc && tsc-alias",
     "watch": "tsc --watch",
-    "test": "go test ./... || echo 'No Go tests found in enclave'",
+    "test": "vitest run",
+    "test:watch": "vitest",
     "lint": "go fmt ./...",
     "clean": "rm -rf dist *.tsbuildinfo"
   },
@@ -39,9 +40,11 @@
     "uint8arrays": "^5.1.0"
   },
   "devDependencies": {
+    "fake-indexeddb": "^6.0.0",
     "tsc-alias": "^1.8.16",
     "typescript": "^5.9.3",
-    "vite": "^7.1.9"
+    "vite": "^7.1.9",
+    "vitest": "^2.1.8"
   },
   "engines": {
     "node": ">=18"
