+++ libs/enclave/src/storage.ts
@@ -1,5 +1,5 @@
 import Dexie, { type Table } from 'dexie';
-import type { StoredVaultState, StoredUCANToken, VaultStorageConfig } from '@sonr.io/enclave';
+import type { StoredVaultState, StoredUCANToken, VaultStorageConfig } from './types.js';
 
 /**
  * Session data stored in IndexedDB
