+++ libs/enclave/vitest.setup.ts
@@ -1,0 +1,2 @@
+// Setup fake IndexedDB for Node.js environment BEFORE any tests run
+import 'fake-indexeddb/auto';
